Change file include lists to work with a makefile
 Changes to be committed:
	modified:   buttonExample.c
	modified:   framebuffer.c
	modified:   main.c
	new file:   makefile
	new file:   shared.h
	modified:   touch.c
The as-forked files had include directives that caused makefile errors, these have now been changed to allow calling a make command to build either the basic drawing
demo using main.c (make draw) or the button demo (make buttontest).  The common includes are now moved to a single header file shared.h.  The draw program works but
the displayed pixels are rotated on the screen by 90 degrees (presumably to a swap of the x and y axis in the QUimat touchscreen?).  The buttontest program has a
major problem, it causese the screen to go white and the display then becomes unusable until you reboot the RPi, however the touch readout is still displayed from
the running buttontest when connected via SSH.